fully factor out science environment
new envs can now be registered and used independently from science
